# cedar-messaging-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe messaging server carries messages to CEDAR users: what the workbench shows as a\nnotification, and what a background process sends when it has something to report.\n\nEvery endpoint acts on the caller's own messages. Sending in a process's name is the one thing\nhere that needs a permission beyond being logged in.\n",
    "title" : "CEDAR Messaging Server API",
    "version" : "2.9.8"
  },
  "paths" : {
...
    "/summary" : {
      "get" : {
        "description" : "Report how many messages the caller has in total, how many are unread, and how many have not yet been notified. This is the counts alone, for a workbench badge that does not want the messages themselves.",
        "operationId" : "getSummary",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageSummary"
                }
              }
            },
            "description" : "Total, unread, and not-yet-notified counts"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Count the caller's messages",
        "tags" : [ "Messages" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Operation that failed. Its fields depend on the operation type.",
              "nullable" : true
            },
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
...
      "MessageSummary" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "notnotified" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unread" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "notnotified", "total", "unread" ]
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}